add: servicios, repository, controller od auht

+++ src/main/java/com/quercus/PineappleSupermarket/repositories/RoleRepository.java
@@ -4,10 +4,10 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.quercus.PineappleSupermarket.enums.ERole;
+import com.quercus.PineappleSupermarket.models.ERole;
 import com.quercus.PineappleSupermarket.models.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 
 	Optional<Role> findByName(ERole name);
 }
